+++ src/Collector.cpp
@@ -71,6 +71,9 @@ void CollectorCharacter::setValue(int i, const Token& t) {
     boost::container::string buffer;
     SourceIterators string = t.getString(&buffer);
 
+    if (t.hasNull())
+      warn(t.row(), t.col(), "", "embedded null");
+
     SET_STRING_ELT(column_, i, pEncoder_->makeSEXP(string.first, string.second));
     break;
   };