+++ src/TokenizerDelim.cpp
@@ -43,12 +43,15 @@ Token TokenizerDelim::nextToken() {
     return Token(TOKEN_EOF, row, col);
 
   SourceIterator token_begin = cur_;
-  bool hasEscapeD = false, hasEscapeB = false;
+  bool hasEscapeD = false, hasEscapeB = false, hasNull = false;
 
   while (cur_ != end_) {
     // Increments cur on destruct, ensuring that we always move on to the
     // next character
     Advance advance(&cur_);
+
+    if (*cur_ == '\0')
+      hasNull = true;
 
     if ((row_ + 1) % 100000 == 0 || (col_ + 1) % 100000 == 0)
       Rcpp::checkUserInterrupt();
@@ -76,15 +79,15 @@ Token TokenizerDelim::nextToken() {
     case STATE_FIELD:
       if (*cur_ == '\r' || *cur_ == '\n') {
         newRecord();
-        return fieldToken(token_begin, advanceForLF(&cur_, end_), hasEscapeB, row, col);
+        return fieldToken(token_begin, advanceForLF(&cur_, end_), hasEscapeB, hasNull, row, col);
       } else if (isComment(cur_)) {
         state_ = STATE_COMMENT;
-        return fieldToken(token_begin, cur_, hasEscapeB, row, col);
+        return fieldToken(token_begin, cur_, hasEscapeB, hasNull, row, col);
       } else if (escapeBackslash_ && *cur_ == '\\') {
         state_ = STATE_ESCAPE_F;
       } else if (*cur_ == delim_) {
         newField();
-        return fieldToken(token_begin, cur_, hasEscapeB, row, col);
+        return fieldToken(token_begin, cur_, hasEscapeB, hasNull, row, col);
       }
       break;
 
@@ -100,15 +103,15 @@ Token TokenizerDelim::nextToken() {
       } else if (*cur_ == '\r' || *cur_ == '\n') {
         newRecord();
         return stringToken(token_begin + 1, advanceForLF(&cur_, end_) - 1,
-          hasEscapeB, hasEscapeD, row, col);
+          hasEscapeB, hasEscapeD, hasNull, row, col);
       } else if (isComment(cur_)) {
         state_ = STATE_COMMENT;
         return stringToken(token_begin + 1, cur_ - 1,
-          hasEscapeB, hasEscapeD, row, col);
+          hasEscapeB, hasEscapeD, hasNull, row, col);
       } else if (*cur_ == delim_) {
         newField();
         return stringToken(token_begin + 1, cur_ - 1,
-          hasEscapeB, hasEscapeD, row, col);
+          hasEscapeB, hasEscapeD, hasNull, row, col);
       } else {
         warn(row, col, "delimiter or quote", std::string(cur_, cur_ + 1));
         state_ = STATE_STRING;
@@ -131,13 +134,13 @@ Token TokenizerDelim::nextToken() {
       if (*cur_ == '\r' || *cur_ == '\n') {
         newRecord();
         return stringToken(token_begin + 1, advanceForLF(&cur_, end_) - 1,
-          hasEscapeB, hasEscapeD, row, col);
+          hasEscapeB, hasEscapeD, hasNull, row, col);
       } else if (isComment(cur_)) {
         state_ = STATE_COMMENT;
-        return stringToken(token_begin + 1, cur_ - 1, hasEscapeB, hasEscapeD, row, col);
+        return stringToken(token_begin + 1, cur_ - 1, hasEscapeB, hasEscapeD, hasNull, row, col);
       } else if (*cur_ == delim_) {
         newField();
-        return stringToken(token_begin + 1, cur_ - 1, hasEscapeB, hasEscapeD, row, col);
+        return stringToken(token_begin + 1, cur_ - 1, hasEscapeB, hasEscapeD, hasNull, row, col);
       } else {
         state_ = STATE_FIELD;
       }
@@ -171,19 +174,19 @@ Token TokenizerDelim::nextToken() {
 
   case STATE_STRING_END:
   case STATE_QUOTE:
-    return stringToken(token_begin + 1, end_ - 1, hasEscapeB, hasEscapeD, row, col);
+    return stringToken(token_begin + 1, end_ - 1, hasEscapeB, hasEscapeD, hasNull, row, col);
 
   case STATE_STRING:
     warn(row, col, "closing quote at end of file");
-    return stringToken(token_begin + 1, end_, hasEscapeB, hasEscapeD, row, col);
+    return stringToken(token_begin + 1, end_, hasEscapeB, hasEscapeD, hasNull, row, col);
 
   case STATE_ESCAPE_S:
   case STATE_ESCAPE_F:
     warn(row, col, "closing escape at end of file");
-    return stringToken(token_begin, end_ - 1, hasEscapeB, hasEscapeD, row, col);
+    return stringToken(token_begin, end_ - 1, hasEscapeB, hasEscapeD, hasNull, row, col);
 
   case STATE_FIELD:
-    return fieldToken(token_begin, end_, hasEscapeB, row, col);
+    return fieldToken(token_begin, end_, hasEscapeB, hasNull, row, col);
 
   case STATE_COMMENT:
     return Token(TOKEN_EOF, row, col);
@@ -215,18 +218,20 @@ Token TokenizerDelim::emptyToken(int row, int col) {
   return Token(TOKEN_EMPTY, row, col).flagNA(NA_);
 }
 
-Token TokenizerDelim::fieldToken(SourceIterator begin, SourceIterator end, bool hasEscapeB,
-  int row, int col)  {
-  Token t(begin, end, row, col, (hasEscapeB) ? this : NULL);
+Token TokenizerDelim::fieldToken(SourceIterator begin, SourceIterator end,
+                                 bool hasEscapeB, bool hasNull,
+                                 int row, int col)  {
+  Token t(begin, end, row, col, hasNull, (hasEscapeB) ? this : NULL);
   if (trimWS_)
     t.trim();
   t.flagNA(NA_);
   return t;
 }
 
-Token TokenizerDelim::stringToken(SourceIterator begin, SourceIterator end, bool hasEscapeB,
-  bool hasEscapeD, int row, int col) {
-  Token t(begin, end, row, col, (hasEscapeD || hasEscapeB) ? this : NULL);
+Token TokenizerDelim::stringToken(SourceIterator begin, SourceIterator end,
+                                  bool hasEscapeB, bool hasEscapeD, bool hasNull,
+                                  int row, int col) {
+  Token t(begin, end, row, col, hasNull, (hasEscapeD || hasEscapeB) ? this : NULL);
   if (trimWS_)
     t.trim();
   return t;