+++ src/TokenizerLine.h
@@ -31,11 +31,16 @@ public:
   Token nextToken() {
     SourceIterator token_begin = cur_;
 
+    bool hasNull = false;
+
     if (!moreTokens_)
       return Token(TOKEN_EOF, line_, 0);
 
     while (cur_ != end_) {
       Advance advance(&cur_);
+
+      if (*cur_ == '\0')
+        hasNull = true;
 
       if ((line_ + 1) % 500000 == 0)
         Rcpp::checkUserInterrupt();
@@ -43,7 +48,7 @@ public:
       switch(*cur_) {
       case '\r':
       case '\n':
-        return Token(token_begin, advanceForLF(&cur_, end_), line_++, 0);
+        return Token(token_begin, advanceForLF(&cur_, end_), hasNull, line_++, 0);
       default:
         break;
       }
@@ -54,7 +59,7 @@ public:
     if (token_begin == end_) {
       return Token(TOKEN_EOF, line_++, 0);
     } else {
-      return Token(token_begin, end_, line_++, 0);
+      return Token(token_begin, end_, hasNull, line_++, 0);
     }
   }